move seek_generator to util
